% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add.cesu.tag}
\alias{add.cesu.tag}
\title{Add a Tag for CESU Member or Not}
\usage{
add.cesu.tag(dat.nspires, cesu = nasaprograminsights::cesu)
}
\arguments{
\item{dat.nspires}{data frame containing proposals from make_nspires}

\item{cesu}{lookup table inherent to pacakge.}
}
\description{
This is in helper because it's of limited value.
}
