% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{analysis_dropdown}
\alias{analysis_dropdown}
\title{Program Analysis Drop Down and Plots UI}
\usage{
tabPanel("Introduction & Setup", analysis_dropdown())
}
\description{
Output drop down menu for user to select a metric to view and the associated calculations and plots for the metric.
}
