% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{analysis_intro}
\alias{analysis_intro}
\title{Analysis Page Intro Text}
\usage{
tabPanel("Program Analysis", analysis_intro())
}
\description{
Output text introducing the user to how to use the drop down menu in the Program Analysis section of the app.
}
