% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{analysisUI}
\alias{analysisUI}
\title{Program Analysis Page UI}
\usage{
ui <- fluidPage(mainPanel(tabsetPanel(analysisUI())))
}
\description{
Base UI function for Program Analysis page calling other UI functions
}
