% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_funs.R
\name{deduplicate_cols}
\alias{deduplicate_cols}
\title{Deduplicate columns in NSPIRES internal dataframes}
\usage{
deduplicate_cols(dat)
}
\arguments{
\item{dat}{a data.frame comprising NSPIRES proposal_master files}
}
\description{
Deduplicate columns in NSPIRES internal dataframes
}
