% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_funs.R
\name{deduplicate_rows}
\alias{deduplicate_rows}
\title{Deduplicate rows (people, proposals) in NSPIRES internal dataframes}
\usage{
deduplicate_rows(dat)
}
\arguments{
\item{dat}{a data.frame comprising NSPIRES either proposal_master or people_master files}
}
\description{
Deduplicate rows (people, proposals) in NSPIRES internal dataframes
}
