% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_setup.R
\name{display_dfs}
\alias{display_dfs}
\title{Display Data Frames UI}
\usage{
tabPanel("Introduction & Setup", display_dfs())
}
\description{
Output drop down menu for user to select a data frame to view, a Verbatim-text output for the user to see how many rows are in the selected data frame,
and a download handler for the user to download the selected data frame.
}
