% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_setup.R
\name{display_dfs_text}
\alias{display_dfs_text}
\title{Display Data Frame Text UI}
\usage{
fluidPage(display_dfs_text())
}
\description{
Output text with instructions for the user to view and download the proposal and people data frames from the app.
}
