% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_setup.R
\name{filter_programs}
\alias{filter_programs}
\title{Filter Programs UI}
\usage{
tabPanel("Introduction & Setup", filter_programs())
}
\description{
Output check box element for user to select program names to filter proposals by and view their selections.
}
