% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffuns.R
\name{get_org_df}
\alias{get_org_df}
\title{Get Org DF}
\usage{
get_org_df(nspires$proposals)
}
\arguments{
\item{df}{data frame containing linked org names}
}
\value{
A data frame containing information about every individual linked organization passed into the function
}
\description{
Function aggregating linked organizations by proposal number.
}
\details{
Similar to the `get_pi_df` function. For each organization, the proposal numbers of the proposals the organization has submitted are stored, the number of proposals it has submitted, the number of proposals that have been selected, and the selection percentage (number of selections / submissions * 100).
}
