% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffuns.R
\name{get_pi_df}
\alias{get_pi_df}
\title{Get PI DF}
\usage{
get_pi_df(nspires$proposals)
}
\arguments{
\item{df}{data frame containing PI SUIDs}
}
\value{
A data frame containing information about every individual PI passed into the function
}
\description{
Function aggregating PI SUIDs by proposal number.
}
\details{
For each PI, the proposal numbers of the proposals the PI has submitted are stored, the number of proposals they have submitted, the number of proposals of theirs that have been selected, and their selection percentage (number of selections / submissions * 100).
}
