% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nspires.R
\name{make_nspires}
\alias{make_nspires}
\title{Make a list of munged NSPIRES data using local data sources.}
\usage{
make_nspires(dir = "data-raw/data-raw-internal/nspires-internal", N = 200)
}
\arguments{
\item{dir}{Directory for where the internal data is stored. This can contain subdirectories, as this function attempts to import data as recursive=TRUE.}

\item{N}{Parameter used to facilitate data import. Can ignore. A # b/w 100-300 is ideal.}
}
\description{
Make a list of munged NSPIRES data using local data sources.
}
