% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungenspiresinternal.R
\name{munge.nspires.internal.peopledata}
\alias{munge.nspires.internal.peopledata}
\title{Munge NSPIRES Internal People Data}
\usage{
munge.nspires.internal.peopledata(df)
}
\arguments{
\item{df}{Data from the internal-access-only NSPIRES. Currently this function only really supports the "people" worksheets.}
}
\description{
returns an object of class data.frame
}
