% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungenspiresinternal.R
\name{munge.nspires.internal.propsdata}
\alias{munge.nspires.internal.propsdata}
\title{Munge NSPIRES Internal Proposal Data}
\usage{
munge.nspires.internal.propsdata(df)
}
\arguments{
\item{df}{Data from the internal-access-only NSPIRES. Currently this function only really supports the "propals" worksheets.}
}
\description{
returns an object of class data.frame
}
