% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_nspires_funs.R
\name{munge.nspires.people}
\alias{munge.nspires.people}
\title{Munge internal NSPIRES people databases}
\usage{
munge.nspires.people(df)
}
\arguments{
\item{df}{Data from the internal-access-only NSPIRES. Currently this function only really supports the "people" worksheets.}
}
\description{
returns an object of class data.frame
}
