% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasaprograminsights-package.R
\docType{package}
\name{nasaprograminsights-package}
\alias{nasaprograminsights}
\alias{nasaprograminsights-package}
\title{nasaprograminsights: Generate Program Insights Using NSPIRES Data}
\description{
A package for NASA personnel to munge, generate simple statistics, visualizations, and summaries of NSPIRES data.
}
\keyword{internal}
