% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_program_name}
\alias{plot_program_name}
\title{Plot Bar Plot by Program Name}
\usage{
plot_program_name(nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ], nspires$proposals[!nspires$proposals$`program name` %in% filtered_programs(), ], "international", "Submitted Proposal Int'l Participation by Program Name")
}
\arguments{
\item{df1}{data frame containing filtered proposals}

\item{df2}{data frame containing other proposals}

\item{metric}{string specifying what the program names should be plotted by}

\item{plot_title}{string specifying the title of the plot}
}
\value{
A plotly object to be printed in the Proposal Analysis tab.
}
\description{
Function plotting a horizontal bar plot grouping a given data frame by program name by a passed in metric. Compares filtered program names to other program names.
}
\details{
The following metrics can be passed into `plot_program_name`:\n
1. "solicitation id"\n
2. "proposal count"\n
3. "selection percentage"\n
4. "total proposed award amount"\n
5. "average proposed award amount"\n
6. "pi"\n
7. "international"\n
8. "nasa"
}
\keyword{internal}
