% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfuns.R
\name{print_org}
\alias{print_org}
\title{Print Linked Organization Statistics}
\usage{
print_org(data.frame(nspires$proposals), data.frame(nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ]))
}
\arguments{
\item{props}{data frame containing all proposals}

\item{filtered_props}{data frame containing only proposals from filtered programs}
}
\value{
A concatenated string to be printed in the Proposal Analysis tab when the `Linked organizations` option is selected in the view drop down.
}
\description{
Function printing information related to all program and filtered programs' linked organizations.
}
\details{
The following information is printed for both all proposals and filtered proposals:\n
1. Number of organizations who submitted a proposal\n
2. Number of organizations who had a selected proposal\n
3. Number of organizations who submitted multiple proposals\n
4. Percentage of organizations who submitted a proposal multiple times\n
5. Number of organizations who were selected multiple times\n
5. Percentage of organizations selected multiple times
}
\keyword{internal}
