% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfuns.R
\name{print_participation}
\alias{print_participation}
\title{Print Proposal Participation Statistics}
\usage{
print_participation(nspires$proposals, nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ])
}
\arguments{
\item{props}{data frame containing all proposals}

\item{filtered_props}{data frame containing only proposals from filtered programs}
}
\value{
A concatenated string to be printed in the Proposal Analysis tab when the `Proposal Participation` option is selected in the view drop down.
}
\description{
Function printing information related to all program and filtered programs' NASA and internatiobal proposal participation.
}
\details{
The following information is printed for both all proposals and filtered proposals:\n
1. Number of submitted proposals with NASA staff participation\n
2. Percentage of submitted proposals with NASA staff participation\n
3. Number of selected proposals with NASA staff participation\n
4. Percentage of selected proposals with NASA staff participation\n
5. Number of submitted proposals with international participation\n
6. Percentage of submitted proposals with international participation\n
7. Number of selected proposals with international participation\n
8. Percentage of selected proposals with international participation\n
}
\keyword{internal}
