% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfuns.R
\name{print_pi}
\alias{print_pi}
\title{Print PI and Member Statistics}
\usage{
print_pi(data.frame(nspires$proposals), data.frame(nspires$people), data.frame(nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ]), data.frame(nspires$people[nspires$people$`program name` %in% filtered_programs(), ]))
}
\arguments{
\item{props}{data frame containing all proposals}

\item{ppl}{data frame containing all people}

\item{filtered_props}{data frame containing only proposals from filtered programs}

\item{filtered_ppl}{data frame containing only people from filtered programs}
}
\value{
A concatenated string to be printed in the Proposal Analysis tab when the `PI and member participation` option is selected in the view drop down.
}
\description{
Function printing information related to all program and filtered programs' pi and member enrollment.
}
\details{
The following information is printed for both all proposals and filtered proposals:\n
1. Number of PIs who submitted a proposal\n
2. Number of PIs who had a selected proposal\n
3. Number of members who were listed on a submitted proposal\n
4. Number of members who were listed on a selected proposal\n
5. Number of members per proposal on average\n
6. The average selection percentage for PIs (number of selections/number of submission for each PI)\n
7. Number of PIs who submitted multiple proposals\n
8. Percentage of PIs involved in multiple submissions\n
9. Number of PIs who were selected multiple times\n
10. Out of PIs who submitted multiple times, percentage of PIs were selected at once.\n
11. The average selection percentage for PIs who submitted multiple times
}
\keyword{internal}
