% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfuns.R
\name{print_proposal_award_amount}
\alias{print_proposal_award_amount}
\title{Print Proposal Award Amount Statistics}
\usage{
print_proposal_award_amount(nspires$proposals, nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ])
}
\arguments{
\item{props}{data frame containing all proposals}

\item{filtered_props}{data frame containing only proposals from filtered programs}
}
\value{
A concatenated string to be printed in the Proposal Analysis tab when the `Proposal award amount` option is selected in the view drop down.
}
\description{
Function printing information related to all program and filtered programs' proposal award amounts.
}
\details{
The following information is printed for both all proposals and filtered proposals:\n
1. The total amount awarded by NASA to selected proposals\n
2. The average amount per proposal awarded by NASA to selected proposals
}
\keyword{internal}
