% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfuns.R
\name{print_proposal_count}
\alias{print_proposal_count}
\title{Print Proposal Count Statistics}
\usage{
print_proposal_count(nspires$proposals, nspires$proposals[nspires$proposals$`program name` %in% filtered_programs(), ])
}
\arguments{
\item{props}{data frame containing all proposals}

\item{filtered_props}{data frame containing only proposals from filtered programs}
}
\value{
A concatenated string to be printed in the Proposal Analysis tab when the `Proposal count` option is selected in the view drop down.
}
\description{
Function printing information related to all program and filtered programs' solicitation and proposal counts.
}
\details{
The following information is printed for both all proposals and filtered proposals:\n
1. Number of solicitations\n
2. Number of submitted proposals\n
3. Number of selected proposals\n
4. Proposal selection percentage (number selected / submitted * 100)
}
