% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runapp}
\alias{runapp}
\title{Build and Open the R Shiny Application}
\usage{
nspires = make_nspires("local file directory with NSPIRES proposals")
runapp(data = nspires, removeNonEA = TRUE)
}
\arguments{
\item{data}{A list containing NSPIRES data. The list should have at least the two following elements (data frames or data tables): c(`people`, `proposals`)}

\item{removeNonEA}{logical Default TRUE will remove any data that is not currently defined as a program under Earth Action.}

\item{apptheme}{ANAND PLS FILL this out --saying its to set the bootswatch theme}
}
\description{
Builds and opens the R Shiny application on local machine. Features include generation and visualization of simple statistics.
}
